♻️ F09: customHooks > useMenu, useBasket, useDisplaySuccessMessage
